@@ -58,8 +33,8 @@ dependencies:
   url_launcher: ^6.1.11
   geolocator: ^13.0.1
   google_maps_flutter: ^2.9.0
-  cloud_firestore: ^5.4.1
-  webview_flutter: ^4.5.0
+  cloud_firestore: ^5.6.3
+  webview_flutter: ^4.10.0
   flutter_facebook_auth: ^6.0.3
   firebase_core: ^3.4.1
   google_sign_in_web: ^0.12.3+2
@@ -73,52 +48,28 @@ dependencies:
   firebase_database: ^11.1.4
   youtube_player_iframe: ^5.2.1
   encrypt: ^5.0.3
-
-
+  payu_checkoutpro_flutter: ^1.3.0
 
 dev_dependencies:
   build_runner: ^2.4.0
   build_web_compilers: ^4.0.4
   flutter_test:
     sdk: flutter
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
   flutter_lints: ^4.0.0
   flutter_launcher_icons: ^0.14.0
+
 flutter_launcher_icons:
-  android: "launcher_icon"
+  android: true
   ios: true
+  remove_alpha_ios: true
   image_path: "assets/icon/app-icon.png"
-  min_sdk_android: 21 # android min sdk min:16, default 21
-  web:
-    generate: true
-    image_path: "assets/icon/app-icon.png"
-    background_color: "#hexcode"
-    theme_color: "#hexcode"
-  windows:
-    generate: true
-    image_path: "assets/icon/app-icon.png"
-    icon_size: 48 # min:48, max:256, default: 48
-  macos:
-    generate: true
-    image_path: "assets/icon/app-icon.png"
+  adaptive_icon_background: "#FFFFFF"
+  adaptive_icon_foreground: "assets/icon/app-icon.png"
+  adaptive_icon_mask: true
+  min_sdk_android: 21
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
     - assets/images/web/
