@@ -45,7 +45,7 @@ class _BalanceState extends State<Balance> {
           IconButton(onPressed: (){
             Navigator.pushAndRemoveUntil(
               context,
-              MaterialPageRoute(builder: (context) => const Dashboard()),
+              MaterialPageRoute(builder: (context) => Dashboard()),
                   (Route<dynamic> route) => false,
             );
           }, icon: const Icon(Icons.home_outlined))
