@@ -42,7 +42,7 @@ class _PayState extends State<Pay> {
           IconButton(onPressed: (){
             Navigator.pushAndRemoveUntil(
               context,
-              MaterialPageRoute(builder: (context) => const Dashboard()),
+              MaterialPageRoute(builder: (context) =>  Dashboard()),
                   (Route<dynamic> route) => false,
             );
           }, icon: const Icon(Icons.home_outlined))
@@ -111,7 +111,7 @@ class _PayState extends State<Pay> {
                             title: LocalizationController.getInstance().getTranslatedValue("Cancel"),buttonHeight: 45, textColor: Colors.white, buttonColors: const [Color(0xFFf2b20a), Color(0xFFf34509)], onPressed: (Offset buttonOffset){
                           Navigator.pushAndRemoveUntil(
                             context,
-                            MaterialPageRoute(builder: (context) => const Dashboard()),
+                            MaterialPageRoute(builder: (context) => Dashboard()),
                                 (Route<dynamic> route) => false,
                           );
                         }),
@@ -120,7 +120,7 @@ class _PayState extends State<Pay> {
                       Expanded(child: GradientButton(
                           title: LocalizationController.getInstance().getTranslatedValue("Send"),buttonHeight: 45, textColor: Colors.white, buttonColors: const [Color(0xFFf2b20a), Color(0xFFf34509)], onPressed: (Offset buttonOffset){
                         print(emailController.text);
-                        paymentController.addOfflineMoney(appLoadController.loggedUserData.value.userid!, emailController.text, amountController.text, appLoadController.loggedUserData.value.token!, context);
+                        // paymentController.addOfflineMoney(appLoadController.loggedUserData.value.userid!, emailController.text, amountController.text, appLoadController.loggedUserData.value.token!, context);
                       }))
                     ],
                   ),
