@@ -106,7 +106,7 @@ class _ViewCommentsState extends State<ViewComments> {
             IconButton(onPressed: (){
               Navigator.pushAndRemoveUntil(
                 context,
-                MaterialPageRoute(builder: (context) => const Dashboard()),
+                MaterialPageRoute(builder: (context) => Dashboard()),
                     (Route<dynamic> route) => false,
               );
             }, icon: const Icon(Icons.home_outlined))
