@@ -89,7 +89,7 @@ class _AddChildInfoState extends State<AddChildInfo> {
           IconButton(onPressed: (){
             Navigator.pushAndRemoveUntil(
               context,
-              MaterialPageRoute(builder: (context) => const Dashboard()),
+              MaterialPageRoute(builder: (context) => Dashboard()),
                   (Route<dynamic> route) => false,
             );
           }, icon: const Icon(Icons.home_outlined))
