@@ -81,7 +81,7 @@ class _AddPrimaryState extends State<AddPrimary> {
           IconButton(onPressed: (){
           Navigator.pushAndRemoveUntil(
           context,
-          MaterialPageRoute(builder: (context) => const Dashboard()),
+          MaterialPageRoute(builder: (context) => Dashboard()),
           (Route<dynamic> route) => false,
           );
           }, icon: const Icon(Icons.home_outlined))
