@@ -138,7 +138,7 @@ class _AddRelativesInfoState extends State<AddRelativesInfo> {
           IconButton(onPressed: (){
             Navigator.pushAndRemoveUntil(
               context,
-              MaterialPageRoute(builder: (context) => const Dashboard()),
+              MaterialPageRoute(builder: (context) => Dashboard()),
                   (Route<dynamic> route) => false,
             );
           }, icon: const Icon(Icons.home_outlined))
@@ -374,7 +374,7 @@ class _AddRelativesInfoState extends State<AddRelativesInfo> {
                       title: LocalizationController.getInstance().getTranslatedValue("Cancel"),buttonHeight: 45, textColor: Colors.white, buttonColors: const [Color(0xFFf2b20a), Color(0xFFf34509)], onPressed: (Offset buttonOffset){
                     Navigator.pushAndRemoveUntil(
                       context,
-                      MaterialPageRoute(builder: (context) => const Dashboard()),
+                      MaterialPageRoute(builder: (context) => Dashboard()),
                           (Route<dynamic> route) => false,
                     );
                   }),
